fix: prevent integer overflow in getInt
It is possible for the raw cast in getInt to return something completely
incorrect due to a long value overflowing the int. Adding some defense
to prevent this.

Also remove the unnecessary boxing and unboxing of longs in getLong.

+++ client/src/com/aerospike/client/Record.java
@@ -95,7 +95,7 @@ public final class Record {
 		// The server always returns numbers as longs if bin found.
 		// If bin not found, the result will be null.  Convert null to zero.
 		Object result = getValue(name);
-		return (result != null)? (Long)result : 0;
+		return (result != null) ? (long) result : 0;
 	}
 
 	/**
@@ -103,7 +103,7 @@ public final class Record {
 	 */
 	public int getInt(String name) {
 		// The server always returns numbers as longs, so get long and cast.
-		return (int)getLong(name);
+		return Math.toIntExact(getLong(name));
 	}
 
 	/**
